Add docs and split files from index.js
