+++ package.json
@@ -10,7 +10,9 @@
     "build": "babel src/ --out-dir lib --source-maps",
     "travis": "npm run coverage",
     "posttravis": "codeclimate-test-reporter < ./coverage/lcov.info",
-    "docs": "rm ./.babelrc && documentation build --github --output docs --format html && git checkout .babelrc && open ./docs/index.html"
+    "predocs": "rm ./.babelrc || true",
+    "docs": "documentation build src/* --github --output docs --format html",
+    "postdocs": " git checkout .babelrc"
   },
   "repository": {
     "type": "git",
