+++ src/isArray.js
@@ -3,5 +3,10 @@
  *
  * @param param Value to be checked
  * @returns {boolean} Returns true when the `param` is an Array
+ *
+ * @example
+ * isArray([]); // Returns: true
+ * isArray({}); // Returns: false
+ * isArray('Hello'); // Returns: false
  */
 export default Array.isArray;
