+++ src/isFunction.js
@@ -2,7 +2,7 @@
  * Check if the value is a Function
  *
  * @name isFunction
- * @param param Value to be checked
+ * @param {*} param Value to be checked
  * @returns {boolean} Returns true when the `param` is a Function
  */
 export {default as default} from 'lodash.isfunction';
