+++ src/isString.js
@@ -1,1 +1,8 @@
+/**
+ * Check if the value is a String
+ *
+ * @name isString
+ * @param param Value to be checked
+ * @returns {boolean} Returns true when the `param` is a String
+ */
 export {default as default} from 'lodash.isstring';
